test(step3): PurgomalumClient 인터페이스 분리
