+++ src/main/java/kitchenpos/infra/DefaultPurgomalumClient.java
@@ -1,0 +1,29 @@
+package kitchenpos.infra;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import java.net.URI;
+
+@Component
+public class DefaultPurgomalumClient implements PurgomalumClient {
+    private final RestTemplate restTemplate;
+    private final ObjectMapper objectMapper;
+
+    public DefaultPurgomalumClient(final RestTemplateBuilder restTemplateBuilder, final ObjectMapper objectMapper) {
+        this.restTemplate = restTemplateBuilder.build();
+        this.objectMapper = objectMapper;
+    }
+
+    @Override
+    public boolean containsProfanity(final String text) {
+        final URI url = UriComponentsBuilder.fromUriString("https://www.purgomalum.com/service/containsprofanity")
+            .queryParam("text", text)
+            .build()
+            .toUri();
+        return Boolean.parseBoolean(restTemplate.getForObject(url, String.class));
+    }
+}
